Move converters to individual files
