+++ dataregistry-extension/src/csvviewer/activate.ts
@@ -6,108 +6,10 @@
  */
 
 import { JupyterFrontEnd } from "@jupyterlab/application";
-import { Registry, createConverter, resolveExtensionConverter } from "@jupyterlab/dataregistry";
-import { datatypes } from "./../datatypes";
-import { mimetypes } from "./../mimetypes";
-import { DataGrid } from "./data_grid";
-
-/**
- * Interface describing an object containing data to convert.
- *
- * @private
- */
-interface Data {
-  /**
-   * Data to convert.
-   */
-  data: any;
-
-  /**
-   * Desired data type.
-   */
-  type: string;
-}
-
-/**
- * Returns a converter for converting from a resolver MIME type to a file MIME type.
- *
- * @private
- * @returns data type converter
- */
-function extension2mimetype() {
-  return resolveExtensionConverter(".csv", mimetypes.csv);
-}
-
-/**
- * Returns a converter for converting text data to CSV.
- *
- * @private
- * @returns data type converter
- */
-function text2csv() {
-  const conversion = {
-    "from": datatypes.text,
-    "to": datatypes.csv
-  };
-
-  return createConverter(conversion, convert);
-
-  /**
-   * Converts from text data to CSV.
-   *
-   * @private
-   * @param obj - data object
-   * @param obj.data - data to convert
-   * @param obj.type - desired type
-   * @returns converted data
-   */
-  function convert(obj: Data) {
-    if (obj.type === mimetypes.csv) {
-      return obj.data;
-    }
-    return null;
-  }
-}
-
-/**
- * Returns a converter for converting CSV data to a data grid.
- *
- * @private
- * @returns data type converter
- */
-function csv2datagrid() {
-  const conversion = {
-    "from": datatypes.csv,
-    "to": datatypes.widget
-  };
-
-  return createConverter(conversion, convert);
-
-  /**
-   * Converts CSV data to a data grid.
-   *
-   * @private
-   * @param obj - data object
-   * @param obj.data - data to convert
-   * @returns converted data
-   */
-  function convert(obj: Data) {
-    return {
-      "type": "Grid",
-      "data": getData
-    };
-
-    /**
-     * Returns a data grid.
-     *
-     * @private
-     * @returns data grid
-     */
-    function getData() {
-      return new DataGrid(obj.data);
-    }
-  }
-}
+import { Registry } from "@jupyterlab/dataregistry";
+import { extension2mimetype } from "./extension_to_mimetype";
+import { text2csv } from "./text_to_csv";
+import { csv2datagrid } from "./csv_to_data_grid";
 
 /**
  * Activates the plugin.
